Fix failing travis builds by dependency pinning
After changing the dependency to Tensorflow<2.0 the ranking tests were
failing, due to small variations in loss values.
To make the tests more robust, we require them to converge to 0 loss now
and thus made the test learning problem much simpler to learn.

This is an overview of the changes:

 * Pin Tensorflow to version 1.x
 * Improve convergence and speed of ranking tests
   - Reduce the number of instances of the learning problem
   - Require learners to reach exactly 0 loss.