Remove usage of Imputer
Imputer was removed from preprocessing in
https://github.com/scikit-learn/scikit-learn/pull/13796. To be
compatible with newer scikit-learn versions, we need to stop using it.
impute.SimpleImputer seems to be a viable replacement.

Do not install scikit-learn thorough conda

It is also installed via pip, leading to conflicts
(https://github.com/scikit-learn/scikit-learn/issues/15858 in
particular).

Make csrank compatible with pymc 3.8

There was a backwards-incompatible change in
https://github.com/pymc-devs/pymc3/commit/521174cfb52a4dfe7b8bd28b1ce193df58e9dd40