Lint python code with flake8
"flake8 is a python tool that glues together pep8, pyflakes, mccabe, and
third-party plugins to check the style and quality of some python code."

Flake8 is reporting a lot of issues to fix. I think its best to get the
checks in first and then gradually tighten the checks as we fix the
outstanding issues.