Declare public interfaces explicitly
As suggested by pep8:

https://www.python.org/dev/peps/pep-0008/#public-and-internal-interfaces

Not doing so trips up flake8's F401, since the imports are not used.
Imports are an implementation detail and should not directly be relied
upon to declare the public API.

Two exceptions to this are csrank/__init__.py and
csrank/dataset_reader/__init__.py, where the API is too big. This will
likely change in the future. See #105.