Determine data dimensionality lazily in FETALinear
Currently we require the data dimensionality to be passed into the
FETALinear constructor. It is then checked that it matches the
dimensionality of the data in `fit`. This is somewhat redundant and not
compatible with the scikit-learn estimator API.

We were already creating the model in the `fit` function. It was also
created when updating the hyperparameters, but that is redundant. In the
`fit` function we can simply derive the dimensionality from the data.