Properly pass on fit kwargs to the core network
Previously we were passing on **kwargs, but also specifying some
arguments explicitly which meant they were not included in **kwargs and
not passed on.

To make this more maintainable, we now also delegate the documentation
to the core.

This changes the test, which is to be expected since the test parameters
(100 epochs instead of 35, no validation instead of 10%) are now
actually respected.