Remove tuning functionality
Since our estimators are at least partially (and will be fully to the
extend possible) compatible with the scikit-learn API now, a custom
solution is no longer necessary. Tuning can be done as for any other
scikit-learn estimator, for example with sklearn.model_selection or with
custom skopt code. In any case, the optimization code does not need to
be in the model library.

Removing this functionality opens the door to potentially reducing
complexity a lot, by removing the "Tunable" class and the
"set_tunable_parameters" functions. These functions have a lot of
duplicated code, and removing them will make further refactorings
easier and likely remove some bugs. Tunable parameters can now be "set"
by simply creating a copy of the learner, since the learners are now
default-constructible and cleanly separate state from configuration.