Remove misc non-parameter attributes from init
Most of these attributes are actually initialized during fit and the
values do not need to be set in init. To conform the the scikit-learn
estimator API, we should not set them in init and postfix values that
are learned from the data with an underscore.

This takes care of most of the remaining attributes in init. The
remaining attributes are due to `construct_model` being called in
`__init__`, which cannot currently be easily fixed. A workaround for
that will follow.