Refactor and document the unit ball sampling
I think default values for internal functions just hinder understanding.
Changed the parameter names to be less domain specific, since we are
just talking about a point in the ball for the purposes of this
function. Since this is an internal function, we can require an already
initialized random state.

Result of this discussion / explanation:
https://github.com/kiudee/cs-ranking/pull/164#discussion_r526701277