Add the pytorch based FATE estimators to the tests
The configuration is based on the configuration of the old (tensorflow
based) fate estimators in the tests. The tensorflow tests used a 10%
validation split, but still verified the performance in-sample. The
validation was not actually used. Therefore I haven't kept that
behavior.

The performance isn't the same. Especially the performance on the choice
task seems worse if we trust the test results. We shouldn't read too
much into that yet. The test is mostly for basic functionality and not a
reliable performance indicator. The sample size is small.