Add a FETA object ranking estimator
These estimators (different tasks, same scoring module) are pretty
repetitive. They are slightly different, so we need different classes.
We could add an additional `FETABasedEstimator` mixin but that would
probably just make it unnecessarily complicated. The current way to
define the estimators also makes it easier to adjust them and play with
different configurations.

The repetitive definition and the additional mixin both have their
advantages. I think we should stick to the repetitive definition for
now.