Add superclasses for skorch based estimators
This is part of the ongoing pytorch migration. We will use skorch as the
basis for our pytorch estimators. That will make it easier to be
compliant with the scikit-learn estimator API.

Ranking and  (general/discrete) choice estimators are often based on
some sort of scoring. The task specific estimators make it easy to
derive concrete estimators from a scoring module.