Add pytorch FATE estimators
This adds a scoring module and the derived estimators for the FATE
approach. The architecture is modular, so it should be easy to
experiment with new ways to put the estimators together. This is a big
commit. Multiple smaller ones that add the separate components (some of
which are structural or can be useful outside of FATE and therefore
could be considered features on their own) would probably have been
better. Splitting it up now would take more time and is not worth it in
this case though.