Unify docs of scoring module based estimators
The documentation of the scoring approach is now consistently delegated
to the docstring of the scoring module. I think that is better, even if
it is a bit harder for the user to find the descripütion. Duplicate
documentation would just get out of sync and increases the friction for
improvements. The class references should make it easy enough to find
the relevant documentation.