Use SELU as the default activation for CmpNet
This is somewhat contrary to the comment I wrote in the general choice
test. I have since noticed a similar issue with RankNet in the tests. It
seems that SELU just works better, at least for these tiny toy problems.
Its already the default for the other pytorch based estimators. I think
it makes sense to use it as the default for CmpNet as well.